pyproject.toml
[project]
name = "deployment_webhook"
version = "0.1.0"
description = ""
authors = [
    {name = "AbooMinister25", email = "aboominister@gmail.com"},
]
dependencies = [
    "bottle>=0.12.25",
    "python-dotenv>=1.0.1",
    "GitPython>=3.1.43",
]
requires-python = ">=3.12"
license = {text = "MIT"}
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.3",
]
format = [
    "isort>=5.10.1",
    "black>=22.8.0",
]
lint = [
    "flake8>=5.0.4",
    "pyright>=1.1.269",
    "flake8-annotations>=2.9.1",
    "flake8-docstrings>=1.6.0",
    "flake8-future-annotations>=0.0.5",
]
repl = [
    "ipython>=7.34.0",
]
coverage = [
    "coverage>=6.4.4",
]
precommit = [
    "pre-commit>=2.20.0",
]

[tool.pdm.scripts]
format = { shell = "black . && isort ." }
lint = { shell = "pyright && flake8" }
test = "pytest ."
coverage = { shell = "coverage run -m pytest && coverage report && coverage xml" }
precommit = { shell = "pdm run pre-commit run --all-files" }
all = { composite = ["format", "lint", "test", "coverage"] }

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pyright]
include = ["deployment_webhook/**/*.py"]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]